реф: работа над структуро проекта
