+++ src/components/app/style.module.scss
@@ -1,0 +1,10 @@
+.container {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+
+  &>div {
+    width: 600px;
+  }
+}
