+++ src/components/burger-ingredients/item.js
@@ -11,13 +11,13 @@ function BurgerIngredientItem({ _id, name, image, proteins, count, onClickHandle
     <div className={`${Styles.item} pl-4 pr-4 mb-10`} data-id={_id} onClick={onClickHandler}>
       {count > 0 && <Counter count={count} size="default" />}
       <img width={252} height={126} alt={name} src={image} />
-      <div className={`${Styles['item-proteins-container']}`}>
-        <div className={`${Styles['item-proteins']} mt-1 mb-1`}>
+      <div className={`${Styles['proteins-container']}`}>
+        <div className={`${Styles['proteins']} mt-1 mb-1`}>
           <span className='text text_type_digits-default'>{proteins}</span>
           <CurrencyIcon />
         </div>
       </div>
-      <div className={`${Styles['item-name-container']}`}>
+      <div className={`${Styles['name']}`}>
         {name}
       </div>
