+++ src/components/order-details/style.module.scss
@@ -1,10 +1,3 @@
-.image {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
 .order-id {
   font-family: 'Iceland';
   font-style: normal;
@@ -18,13 +11,4 @@
 
 .sub-title {
   text-align: center;
-}
-
-.details {
-  display: flex;
-  text-align: center;
-  justify-content: space-around;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
 }
